OverrideSignatureAnalyzer as plugin
